separated spec files for team member
