+++ cypress/integration/v2/blocks/team-member/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Team Member Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
